+++ cmake/Modules/SelectHashes.cmake
@@ -30,7 +30,7 @@ IF(SHA1_BACKEND STREQUAL "CollisionDetection")
 	ADD_DEFINITIONS(-DSHA1DC_NO_STANDARD_INCLUDES=1)
 	ADD_DEFINITIONS(-DSHA1DC_CUSTOM_INCLUDE_SHA1_C=\"common.h\")
 	ADD_DEFINITIONS(-DSHA1DC_CUSTOM_INCLUDE_UBC_CHECK_C=\"common.h\")
-	FILE(GLOB SRC_SHA1 hash/hash_collisiondetect.c hash/sha1dc/*)
+	FILE(GLOB SRC_SHA1 hash/sha1/collisiondetect.c hash/sha1/sha1dc/*)
 ELSEIF(SHA1_BACKEND STREQUAL "OpenSSL")
 	# OPENSSL_FOUND should already be set, we're checking HTTPS_BACKEND
 
@@ -40,13 +40,13 @@ ELSEIF(SHA1_BACKEND STREQUAL "OpenSSL")
 	ELSE()
 		LIST(APPEND LIBGIT2_PC_REQUIRES "openssl")
 	ENDIF()
-	FILE(GLOB SRC_SHA1 hash/hash_openssl.c)
+	FILE(GLOB SRC_SHA1 hash/sha1/openssl.c)
 ELSEIF(SHA1_BACKEND STREQUAL "CommonCrypto")
 	SET(GIT_SHA1_COMMON_CRYPTO 1)
-	FILE(GLOB SRC_SHA1 hash/hash_common_crypto.c)
+	FILE(GLOB SRC_SHA1 hash/sha1/common_crypto.c)
 ELSEIF(SHA1_BACKEND STREQUAL "mbedTLS")
 	SET(GIT_SHA1_MBEDTLS 1)
-	FILE(GLOB SRC_SHA1 hash/hash_mbedtls.c)
+	FILE(GLOB SRC_SHA1 hash/sha1/mbedtls.c)
 	LIST(APPEND LIBGIT2_SYSTEM_INCLUDES ${MBEDTLS_INCLUDE_DIR})
 	LIST(APPEND LIBGIT2_LIBS ${MBEDTLS_LIBRARIES})
 	# mbedTLS has no pkgconfig file, hence we can't require it
@@ -55,9 +55,9 @@ ELSEIF(SHA1_BACKEND STREQUAL "mbedTLS")
 	LIST(APPEND LIBGIT2_PC_LIBS ${MBEDTLS_LIBRARIES})
 ELSEIF(SHA1_BACKEND STREQUAL "Win32")
 	SET(GIT_SHA1_WIN32 1)
-	FILE(GLOB SRC_SHA1 hash/hash_win32.c)
+	FILE(GLOB SRC_SHA1 hash/sha1/win32.c)
 ELSEIF(SHA1_BACKEND STREQUAL "Generic")
-	FILE(GLOB SRC_SHA1 hash/hash_generic.c)
+	FILE(GLOB SRC_SHA1 hash/sha1/generic.c)
 # ELSEIF(NOT USE_SHA1)
 ELSE()
 	MESSAGE(FATAL_ERROR "Asked for unknown SHA1 backend: ${SHA1_BACKEND}")
