+++ src/hash/sha1/collisiondetect.h
@@ -5,8 +5,8 @@
  * a Linking Exception. For full terms see the included COPYING file.
  */
 
-#ifndef INCLUDE_hash_hash_collisiondetect_h__
-#define INCLUDE_hash_hash_collisiondetect_h__
+#ifndef INCLUDE_hash_sha1_collisiondetect_h__
+#define INCLUDE_hash_sha1_collisiondetect_h__
 
 #include "hash.h"
