+++ src/hash/sha1/mbedtls.c
@@ -5,7 +5,7 @@
  * a Linking Exception. For full terms see the included COPYING file.
  */
 
-#include "hash_openssl.h"
+#include "mbedtls.h"
 
 int git_hash_global_init(void)
 {
@@ -19,41 +19,28 @@ int git_hash_ctx_init(git_hash_ctx *ctx)
 
 void git_hash_ctx_cleanup(git_hash_ctx *ctx)
 {
-	GIT_UNUSED(ctx);
+    assert(ctx);
+    mbedtls_sha1_free(&ctx->c);
 }
 
 int git_hash_init(git_hash_ctx *ctx)
 {
-	assert(ctx);
-
-	if (SHA1_Init(&ctx->c) != 1) {
-		git_error_set(GIT_ERROR_SHA1, "hash_openssl: failed to initialize hash context");
-		return -1;
-	}
-
-	return 0;
+    assert(ctx);
+    mbedtls_sha1_init(&ctx->c);
+    mbedtls_sha1_starts(&ctx->c);
+    return 0;
 }
 
 int git_hash_update(git_hash_ctx *ctx, const void *data, size_t len)
 {
-	assert(ctx);
-
-	if (SHA1_Update(&ctx->c, data, len) != 1) {
-		git_error_set(GIT_ERROR_SHA1, "hash_openssl: failed to update hash");
-		return -1;
-	}
-
-	return 0;
+    assert(ctx);
+    mbedtls_sha1_update(&ctx->c, data, len);
+    return 0;
 }
 
 int git_hash_final(git_oid *out, git_hash_ctx *ctx)
 {
-	assert(ctx);
-
-	if (SHA1_Final(out->id, &ctx->c) != 1) {
-		git_error_set(GIT_ERROR_SHA1, "hash_openssl: failed to finalize hash");
-		return -1;
-	}
-
-	return 0;
+    assert(ctx);
+    mbedtls_sha1_finish(&ctx->c, out->id);
+    return 0;
 }
