+++ src/hash/sha1/mbedtls.h
@@ -5,15 +5,15 @@
  * a Linking Exception. For full terms see the included COPYING file.
  */
 
-#ifndef INCLUDE_hash_hash_generic_h__
-#define INCLUDE_hash_hash_generic_h__
+#ifndef INCLUDE_hash_sha1_mbedtls_h__
+#define INCLUDE_hash_sha1_mbedtls_h__
 
 #include "hash.h"
 
+#include <mbedtls/sha1.h>
+
 struct git_hash_ctx {
-	unsigned long long size;
-	unsigned int H[5];
-	unsigned int W[16];
+    mbedtls_sha1_context c;
 };
 
-#endif
+#endif /* INCLUDE_hash_sha1_mbedtls_h__ */
