+++ src/hash/sha1/openssl.c
@@ -5,7 +5,7 @@
  * a Linking Exception. For full terms see the included COPYING file.
  */
 
-#include "hash_collisiondetect.h"
+#include "openssl.h"
 
 int git_hash_global_init(void)
 {
@@ -25,22 +25,33 @@ void git_hash_ctx_cleanup(git_hash_ctx *ctx)
 int git_hash_init(git_hash_ctx *ctx)
 {
 	assert(ctx);
-	SHA1DCInit(&ctx->c);
+
+	if (SHA1_Init(&ctx->c) != 1) {
+		git_error_set(GIT_ERROR_SHA1, "hash_openssl: failed to initialize hash context");
+		return -1;
+	}
+
 	return 0;
 }
 
 int git_hash_update(git_hash_ctx *ctx, const void *data, size_t len)
 {
 	assert(ctx);
-	SHA1DCUpdate(&ctx->c, data, len);
+
+	if (SHA1_Update(&ctx->c, data, len) != 1) {
+		git_error_set(GIT_ERROR_SHA1, "hash_openssl: failed to update hash");
+		return -1;
+	}
+
 	return 0;
 }
 
 int git_hash_final(git_oid *out, git_hash_ctx *ctx)
 {
 	assert(ctx);
-	if (SHA1DCFinal(out->id, &ctx->c)) {
-		git_error_set(GIT_ERROR_SHA1, "SHA1 collision attack detected");
+
+	if (SHA1_Final(out->id, &ctx->c) != 1) {
+		git_error_set(GIT_ERROR_SHA1, "hash_openssl: failed to finalize hash");
 		return -1;
 	}
