+++ src/hash/sha1/openssl.h
@@ -5,8 +5,8 @@
  * a Linking Exception. For full terms see the included COPYING file.
  */
 
-#ifndef INCLUDE_hash_hash_openssl_h__
-#define INCLUDE_hash_hash_openssl_h__
+#ifndef INCLUDE_hash_sha1_openssl_h__
+#define INCLUDE_hash_sha1_openssl_h__
 
 #include "common.h"
