+++ src/hash/sha1/sha1dc/sha1.c
