+++ src/hash/sha1/win32.h
